requirements.txt
python-dotenv
Flask==3.1.1
google==3.0.0
google-genai==1.27.0
google-generativeai==0.8.5
